backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
supabase==2.9.1
psycopg2-binary==2.9.10

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
bcrypt==4.2.1

# Environment & Config
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Date/Time
python-dateutil==2.9.0

# Utilities
email-validator==2.2.0
gspread==6.1.2
google-auth==2.35.0
